--- client/src/main/java/com/sparta/rooibos/client/application/dto/condition/SearchClientApplicationCondition.java
@@ -1,3 +1,0 @@
-package com.sparta.rooibos.client.application.dto.condition;
-
-public record SearchClientApplicationCondition(String name,String address,String type,String deleteCheck, String sort, Integer page, Integer size) {}
