+++ client/src/main/java/com/sparta/rooibos/client/application/dto/req/CreateClientApplicationRequest.java
@@ -1,10 +1,3 @@
 package com.sparta.rooibos.client.application.dto.req;
 
-import com.sparta.rooibos.client.presentation.dto.req.CreateClientRequest;
-
-//TODO
-public record CreateClientApplicationRequest(String name, String clientType, String managedHubId, String address) {
-    public CreateClientApplicationRequest(CreateClientRequest createClientRequest) {
-        this(createClientRequest.name(), createClientRequest.clientType(), createClientRequest.managedHubId(), createClientRequest.address());
-    }
-}
+public record CreateClientApplicationRequest(String name, String clientType, String managedHubId, String address) {}
