+++ client/src/main/java/com/sparta/rooibos/client/application/dto/req/SearchClientApplicationRequest.java
@@ -2,7 +2,4 @@ package com.sparta.rooibos.client.application.dto.req;
 
 import org.springframework.data.domain.Pageable;
 
-public record SearchClientApplicationRequest(String name, String address, String type, Boolean isDeleted, Pageable pageable) {
-
-
-}
+public record SearchClientApplicationRequest(String name, String address, String type, Boolean isDeleted, Pageable pageable) {}
