+++ client/src/main/java/com/sparta/rooibos/client/application/dto/req/UpdateClientApplicationRequest.java
@@ -2,6 +2,4 @@ package com.sparta.rooibos.client.application.dto.req;
 
 import java.util.UUID;
 
-public record UpdateClientApplicationRequest(UUID clientId, String name, String address) {
-
-}
+public record UpdateClientApplicationRequest(UUID clientId, String name, String address) {}
