+++ client/src/main/java/com/sparta/rooibos/client/application/dto/req/UpdateHubIdApplicationRequest.java
@@ -2,5 +2,4 @@ package com.sparta.rooibos.client.application.dto.req;
 
 import java.util.UUID;
 
-public record UpdateHubIdApplicationRequest(UUID clientId, String hubId) {
-}
+public record UpdateHubIdApplicationRequest(UUID clientId, String hubId) {}
