+++ client/src/main/java/com/sparta/rooibos/client/application/dto/res/SearchClientApplicationListResponse.java
@@ -2,5 +2,4 @@ package com.sparta.rooibos.client.application.dto.res;
 
 import java.util.UUID;
 
-public record SearchClientApplicationListResponse(UUID id, String name, String type, String managedHubId, String address, boolean isDeleted) {
-}
+public record SearchClientApplicationListResponse(UUID id, String name, String type, String managedHubId, String address, boolean isDeleted) {}
