+++ client/src/main/java/com/sparta/rooibos/client/application/dto/res/SearchClientApplicationResponse.java
@@ -3,6 +3,4 @@ package com.sparta.rooibos.client.application.dto.res;
 import java.util.List;
 
 public record SearchClientApplicationResponse(List<SearchClientApplicationListResponse> clients, long totalCount,
-                                              long page, long size) {
-
-}
+                                              long page, long size) {}
