+++ client/src/main/java/com/sparta/rooibos/client/infrastructure/repository/JpaClientRepository.java
@@ -6,5 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
 
-public interface JpaClientRepository extends JpaRepository<Client, UUID>, ClientRepository {
-}
+public interface JpaClientRepository extends JpaRepository<Client, UUID>, ClientRepository {}
